+++ src/components/store/demo.ts
@@ -1,4 +1,4 @@
-import { ActionType } from './types/ActionType';
+import { ActionType } from '../../../types/store/ActionType';
 import store from './index';
 
 /**
